ISSUE #3307 - make navigation tabs styling reusable
