+++ frontend/src/v5/ui/components/shared/navigationTabs/index.ts
@@ -1,5 +1,5 @@
 /**
- *  Copyright (C) 2021 3D Repo Ltd
+ *  Copyright (C) 2022 3D Repo Ltd
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU Affero General Public License as
@@ -14,11 +14,6 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import { ProjectNavigation } from '@components/shared/projectNavigation';
-import { Wrapper } from './header.styles';
 
-export const Header = (): JSX.Element => (
-	<Wrapper>
-		<ProjectNavigation />
-	</Wrapper>
-);
+export { ProjectNavigation } from './projectNavigation/projectNavigation.component';
+export { TeamspaceNavigation } from './teamspaceNavigation/TeamspaceNavigation.component';
