+++ frontend/src/v5/ui/components/shared/navigationTabs/navigationTabs.styles.ts
